% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_functional_clustering.R
\name{bifun_clu}
\alias{bifun_clu}
\title{main function for bifunctional clustering}
\usage{
bifun_clu(
  data1,
  data2,
  k,
  Time1 = NULL,
  Time2 = NULL,
  trans = log10,
  inv.cov = NULL,
  initial.pars = NULL,
  iter.max = 100,
  parscale = 0.001
)
}
\arguments{
\item{data1}{matrix or data for cluster}

\item{data2}{matrix or data for cluster}

\item{k}{vector for the cluster number}

\item{Time1}{vector for the time point}

\item{Time2}{vector for the time point}

\item{trans}{indicate log/log2/log10 transform dataset}

\item{inv.cov}{matrix for directly solve cov matrix, default not given(currently not available)}

\item{initial.pars}{vector for manual give initial parameters, default not given}

\item{iter.max}{scales control iteration for EM algorithm}

\item{parscale}{scales control parameters scales for cov pars}
}
\value{
the initial parameters for functional clustering
}
\description{
main function for bifunctional clustering
}
