% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_functional_clustering.R
\name{bifun_clu_parallel}
\alias{bifun_clu_parallel}
\title{parallel version for functional clustering}
\usage{
bifun_clu_parallel(
  data1,
  data2,
  Time1 = NULL,
  Time2 = NULL,
  trans = log10,
  start,
  end,
  iter.max = 100,
  thread = 2
)
}
\arguments{
\item{data1}{data for cluster}

\item{data2}{data for cluster}

\item{Time1}{vector for the time point}

\item{Time2}{vector for the time point}

\item{trans}{indicate log/log2/log10 transform dataset}

\item{start}{vector for the minimum cluster number}

\item{end}{vector for the maximum cluster number}

\item{iter.max}{scales control iteration for EM algorithm}

\item{thread}{scales for how many thread used}
}
\value{
the initial parameters for functional clustering
}
\description{
parallel version for functional clustering
}
