% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_functional_clustering.R
\name{bifun_clu_plot}
\alias{bifun_clu_plot}
\title{bifunctional clustering plot}
\usage{
bifun_clu_plot(result, best.k, label = math_format(expr = 10^.x), n1, n2)
}
\arguments{
\item{result}{list directly from bifun_clu_parallel function}

\item{best.k}{scale of BIC-determined cluster number}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{n1}{scalar for number of column contain first trait/location etc}

\item{n2}{scalar for number of column contain second trait/location etc}
}
\value{
functional clustering plot
}
\description{
bifunctional clustering plot
}
