% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_functional_clustering.R
\name{biget_par_int}
\alias{biget_par_int}
\title{acquire initial parameters for functional clustering}
\usage{
biget_par_int(X, k, times1, times2, n1, n2)
}
\arguments{
\item{X}{matrix for cluster}

\item{k}{vector for the cluster number}

\item{times1}{vector for the x values or time points}

\item{times2}{vector for the x values or time points}

\item{n1}{scalar for number of column contain first trait/location etc}

\item{n2}{scalar for number of column contain second trait/location etc}
}
\value{
the initial parameters for functional clustering
}
\description{
acquire initial parameters for functional clustering
}
