% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{bipower_equation_plot}
\alias{bipower_equation_plot}
\title{plot power equation fitting results for bi-variate model}
\usage{
bipower_equation_plot(
  result,
  label = 10,
  n = 9,
  show.legend = FALSE,
  color1 = "#38E54D",
  color2 = "#FF8787"
)
}
\arguments{
\item{result}{list object returned from data_match}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{n}{scales for how many subplots needed}

\item{show.legend}{show legend or not}

\item{color1}{Hex Color Codes for first data}

\item{color2}{Hex Color Codes for second data}
}
\value{
plot show power curve fitting result
}
\description{
plot power equation fitting results for bi-variate model
}
