% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_functional_clustering.R
\name{biQ_function}
\alias{biQ_function}
\title{Q-function to replace log-likelihood function}
\usage{
biQ_function(par, prob_log, omega_log, X, k, n1, n2, times1, times2)
}
\arguments{
\item{par}{numeric vector for parameters need to be estimated}

\item{prob_log}{mixture component weights(log)}

\item{omega_log}{latent variables(log)}

\item{X}{matrix for cluster}

\item{k}{vector for the cluster number}

\item{n1}{scalar for number of column contain first trait/location etc}

\item{n2}{scalar for number of column contain second trait/location etc}

\item{times1}{vector for the x values or time points}

\item{times2}{vector for the x values or time points}
}
\value{
the Loglikelihood value
}
\description{
Q-function to replace log-likelihood function
}
