% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{biqdODE_plot_all}
\alias{biqdODE_plot_all}
\title{plot all decompose plot for two data}
\usage{
biqdODE_plot_all(
  result1,
  result2,
  label = 10,
  show.legend = FALSE,
  remove.label = TRUE,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{result1}{list of qdODE all for first data}

\item{result2}{list of qdODE all for second data}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{show.legend}{to show legend}

\item{remove.label}{to remove x and y label}

\item{nrow}{scalar for subplot row number}

\item{ncol}{scalar for subplot column number}
}
\description{
plot all decompose plot for two data
}
