% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{darken}
\alias{darken}
\title{make color more dark}
\usage{
darken(color, factor = 1.2)
}
\arguments{
\item{color}{hex color code}

\item{factor}{scalar for darken level}
}
\value{
darkened hex color code
}
\description{
make color more dark
}
\examples{
darken("#FF0000")
}
