% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{data_cleaning}
\alias{data_cleaning}
\title{remove observation with too many 0 values}
\usage{
data_cleaning(data, x = round(ncol(data) * 0.3))
}
\arguments{
\item{data}{dataframe of imported dataset, must have first column as ID}

\item{x}{scales indicate how many 0 to remove}
}
\value{
a dataframe without too many 0 observations
}
\description{
remove observation with too many 0 values
}
\examples{
data_cleaning(matrix(c(c(0,1,1,0,0,1,1), c(2,1,0,3,5,2,2), c(1,1,3,2,4,5,1)), 3, 7), 2)
}
