% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{data_match}
\alias{data_match}
\title{match power_equation fit result for bi-variate model}
\usage{
data_match(result1, result2)
}
\arguments{
\item{result1}{list object from power_equation fit}

\item{result2}{list object from power_equation fit}
}
\value{
a id match list for input dataset
}
\description{
match power_equation fit result for bi-variate model
}
