% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{fun_clu}
\alias{fun_clu}
\title{main function for functional clustering}
\usage{
fun_clu(
  data,
  k,
  Time = NULL,
  trans = log10,
  inv.cov = NULL,
  initial.pars = NULL,
  iter.max = 100,
  parscale = 0.1
)
}
\arguments{
\item{data}{matrix or data for cluster}

\item{k}{vector for the cluster number}

\item{Time}{vector for the time point}

\item{trans}{indicate log/log2/log10 transform dataset}

\item{inv.cov}{matrix for directly solve cov matrix, default not given(currently not available)}

\item{initial.pars}{vector for manual give initial parameters, default not given}

\item{iter.max}{scales control iteration for EM algorithm}

\item{parscale}{scales control parameters scales for cov pars}
}
\value{
the initial parameters for functional clustering
}
\description{
main function for functional clustering
}
