% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{fun_clu_convert}
\alias{fun_clu_convert}
\title{convert result of functional clustering result}
\usage{
fun_clu_convert(result, best.k)
}
\arguments{
\item{result}{list directly from fun_clu_parallel function}

\item{best.k}{scale of BIC-determined cluster number}
}
\value{
list contain module data and fitted data
}
\description{
convert result of functional clustering result
}
