% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{fun_clu_parallel}
\alias{fun_clu_parallel}
\title{parallel version for functional clustering}
\usage{
fun_clu_parallel(
  data,
  Time = NULL,
  trans = log10,
  start,
  end,
  iter.max = 100,
  thread = 2
)
}
\arguments{
\item{data}{data for cluster}

\item{Time}{vector for the time point}

\item{trans}{indicate log/log2/log10 transform dataset}

\item{start}{vector for the minimum cluster number}

\item{end}{vector for the maximum cluster number}

\item{iter.max}{scales control iteration for EM algorithm}

\item{thread}{scales for how many threads used}
}
\value{
the initial parameters for functional clustering
}
\description{
parallel version for functional clustering
}
