% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{fun_clu_plot}
\alias{fun_clu_plot}
\title{functional clustering plot}
\usage{
fun_clu_plot(result, best.k, label = math_format(expr = 10^.x))
}
\arguments{
\item{result}{list directly from fun_clu_parallel function}

\item{best.k}{scale of BIC-determined cluster number}

\item{label}{relabel x and y label due to log-transform, set 10 as default}
}
\value{
functional clustering plot
}
\description{
functional clustering plot
}
