% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{fun_clu_select}
\alias{fun_clu_select}
\title{select result of functional clustering result}
\usage{
fun_clu_select(result_fit, result_funclu, i)
}
\arguments{
\item{result_fit}{list directly from power_equation_fit}

\item{result_funclu}{list from fun_clu_convert}

\item{i}{scale of which cluster selected}
}
\value{
list contain microbe data and fitted data
}
\description{
select result of functional clustering result
}
