% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_biSAD1}
\alias{get_biSAD1}
\title{generate biSAD1 covariance matrix}
\usage{
get_biSAD1(par, n1, n2)
}
\arguments{
\item{par}{vector with four number, first two for ck and the rest for stress}

\item{n1}{scalar indicate length of time1}

\item{n2}{scalar indicate length of time2}
}
\value{
biSAD1 covariance matrix
}
\description{
generate biSAD1 covariance matrix
}
\examples{
get_biSAD1(par=c(2,0.5,2,0.1),n1=4, n2 = 5)
}
