% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{get_interaction}
\alias{get_interaction}
\title{Lasso-based variable selection}
\usage{
get_interaction(data, col, reduction = FALSE)
}
\arguments{
\item{data}{data of clustered results, do not contain cluster column}

\item{col}{scalar of which row number selected}

\item{reduction}{use n/log(n) dimension reduction}
}
\value{
list contain relationship of each row
}
\description{
Lasso-based variable selection
}
