% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_legendre_par}
\alias{get_legendre_par}
\title{use legendre polynomials to fit a given data}
\usage{
get_legendre_par(y, legendre_order, x)
}
\arguments{
\item{y}{vector equal to the y observed data(in this case generic effect)}

\item{legendre_order}{scalar of legendre polynomials}

\item{x}{vector equal to the x value for legendre polynomials(in this case times)}
}
\value{
the polynomials coefficients
}
\description{
use legendre polynomials to fit a given data
}
\examples{
get_legendre_par(14:1,4,1:14)
}
