% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_mu}
\alias{get_mu}
\title{curve fit with modified logistic function}
\usage{
get_mu(mu_par, times)
}
\arguments{
\item{mu_par}{vector with five number}

\item{times}{vector of time point}
}
\value{
numeric vector with the same length to times
}
\description{
curve fit with modified logistic function
}
\examples{
get_mu(mu_par = 1:5, times = 1:14)
}
