% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_mu2}
\alias{get_mu2}
\title{generate mean vectors with ck and stress condition}
\usage{
get_mu2(par, times)
}
\arguments{
\item{par}{vector with ten number, first five for ck and the rest for stress}

\item{times}{vector of time point}
}
\value{
numeric vector with the double length to times
}
\description{
generate mean vectors with ck and stress condition
}
\examples{
get_mu2(par = 1:10, times = 1:14)
}
