% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{get_par_int}
\alias{get_par_int}
\title{acquire initial parameters for functional clustering}
\usage{
get_par_int(X, k, times)
}
\arguments{
\item{X}{matrix for cluster}

\item{k}{vector for the cluster number}

\item{times}{vector for the x values or time points}
}
\value{
the initial parameters for functional clustering
}
\description{
acquire initial parameters for functional clustering
}
