% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_SAD1_covmatrix}
\alias{get_SAD1_covmatrix}
\title{generate standard SAD1 covariance matrix}
\usage{
get_SAD1_covmatrix(par, n)
}
\arguments{
\item{par}{vector with two number for SAD1 covariance matrix}

\item{n}{scalar indicate length of time d}
}
\value{
SAD1 covariance matrix
}
\description{
generate standard SAD1 covariance matrix
}
\examples{
get_SAD1_covmatrix(par = c(2,0.5), n = 14)
}
