% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_microbe}
\alias{gut_microbe}
\title{gut microbe OTU data (species level)}
\format{
A data frame with 65 rows and 21 column, contain first column as microbe ID:
}
\usage{
data(gut_microbe)
}
\description{
The original nucleotide sequences of this study were deposited to the NCBI
Sequence Read Archive under accession number SRP128619.
}
\keyword{datasets}
