% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{logsumexp}
\alias{logsumexp}
\title{calculate log-sum-exp values}
\usage{
logsumexp(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
log-sum-exp values
}
\description{
calculate log-sum-exp values
}
\examples{
logsumexp(c(100, 1000, 10000))
}
