% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mustard_microbe}
\alias{mustard_microbe}
\title{mustard microbe OTU data}
\format{
A data frame with 633 rows and 11 column, contain first column as microbe ID:
}
\usage{
data(mustard_microbe)
}
\description{
Wagner, M. R. et al. Host genotype and age shape the leaf and root microbiomes of a wild perennial plant.
Nat. Commun. 7:12151 doi: 10.1038/ncomms12151 (2016)
}
\keyword{datasets}
