% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction.R
\name{network_maxeffect}
\alias{network_maxeffect}
\title{convert ODE results(ODE_solving2) to basic network plot table}
\usage{
network_maxeffect(result)
}
\arguments{
\item{result}{list result from qsODE_parallel}
}
\value{
a list with basic information to plot network
}
\description{
convert ODE results(ODE_solving2) to basic network plot table
}
