% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{normalization}
\alias{normalization}
\title{min-max normalization}
\usage{
normalization(x, z = 0.2)
}
\arguments{
\item{x}{numeric vector}

\item{z}{scalar add minimum value to avoid 0}
}
\value{
normalized vector
}
\description{
min-max normalization
}
\examples{
normalization(runif(100,min = -100, max = 100))
}
