% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{power_equation}
\alias{power_equation}
\title{use power equation parameters to generate y values}
\usage{
power_equation(x, power_par)
}
\arguments{
\item{x}{vector for x values}

\item{power_par}{matrix contain parameters for power equation}
}
\value{
y values for given power equation parameters
}
\description{
use power equation parameters to generate y values
}
\examples{
power_equation(c(1,2,3,5,7), matrix(c(2,1,1,2),2,2))
}
