% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{power_equation_all}
\alias{power_equation_all}
\title{use power equation to fit observed values}
\usage{
power_equation_all(x, y, maxit = 100)
}
\arguments{
\item{x}{vector for x values}

\item{y}{vector for y values}

\item{maxit}{numeric value for maximum initial pars try}
}
\value{
nls model
}
\description{
use power equation to fit observed values
}
\examples{
power_equation_all(c(1,2,3,5,7), c(5,10,15,17,20))
}
