% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{power_equation_fit}
\alias{power_equation_fit}
\title{use power equation to fit given dataset}
\usage{
power_equation_fit(data, n = 30, trans = log10, thread = 2)
}
\arguments{
\item{data}{cleaned dataframe}

\item{n}{scales for how many interpolation needed}

\item{trans}{indicate log/log2/log10 transform dataset}

\item{thread}{scales for how many thread used}
}
\value{
list contain power equation parameters and fitted data
}
\description{
use power equation to fit given dataset
}
