% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_equation_fitting.R
\name{power_equation_plot}
\alias{power_equation_plot}
\title{plot power equation fitting results}
\usage{
power_equation_plot(result, label = 10, n = 9)
}
\arguments{
\item{result}{list object returned from power_equation_fit}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{n}{scales for how many subplots needed}
}
\value{
plot show power curve fitting result
}
\description{
plot power equation fitting results
}
