% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering.R
\name{Q_function}
\alias{Q_function}
\title{Q-function to replace log-likelihood function}
\usage{
Q_function(par, prob_log, omega_log, X, k, times)
}
\arguments{
\item{par}{numeric vector for parameters need to be estimated}

\item{prob_log}{mixture component weights(log)}

\item{omega_log}{latent variables(log)}

\item{X}{matrix for cluster}

\item{k}{vector for the cluster number}

\item{times}{vector for the x values or time points}
}
\value{
the Loglikelihood value
}
\description{
Q-function to replace log-likelihood function
}
