% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_all}
\alias{qdODE_all}
\title{wrapper for qdODE model}
\usage{
qdODE_all(
  result,
  relationship,
  i,
  init_pars = 1,
  LOP_order = 6,
  method = "ls",
  new_time = NULL,
  n_expand = 100,
  maxit = 1000
)
}
\arguments{
\item{result}{result from power_equation_fit}

\item{relationship}{list contain variable selection results}

\item{i}{scalar for which id used for qdODE solving, must <= nrow}

\item{init_pars}{scalar for initial parameters}

\item{LOP_order}{scalar of LOP order}

\item{method}{scalar of qdODE solving methodm, cuurent only support least square}

\item{new_time}{vector produce new defined time point}

\item{n_expand}{scalar for how many interpolation needed}

\item{maxit}{scalar of Optim iteration setting}
}
\value{
list contain variable selection results and LOP parameters for every row
}
\description{
wrapper for qdODE model
}
