% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_fit}
\alias{qdODE_fit}
\title{legendre polynomials fit to qdODE model}
\usage{
qdODE_fit(
  pars,
  data,
  Time,
  power_par,
  LOP_order = 6,
  new_time = NULL,
  n_expand = 100
)
}
\arguments{
\item{pars}{vector of qdODE parameters}

\item{data}{dataframe of observed data}

\item{Time}{vector of time point}

\item{power_par}{matrix of power equation parameters for dependent effect}

\item{LOP_order}{scalar of LOP order}

\item{new_time}{vector produce new defined time point}

\item{n_expand}{scalar for how many interpolation needed}
}
\value{
list contain legendre polynomials parameters, qdODE values and LOP fitted values
}
\description{
legendre polynomials fit to qdODE model
}
