% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_ls}
\alias{qdODE_ls}
\title{least-square fit for qdODE model}
\usage{
qdODE_ls(pars, data, Time, power_par)
}
\arguments{
\item{pars}{vector for unknown ODE parameters}

\item{data}{data contain independent effect as first row and dependent effect}

\item{Time}{vector of time point}

\item{power_par}{matrix of power equation parameters for dependent effect}
}
\value{
mean-square error
}
\description{
least-square fit for qdODE model
}
