% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_parallel}
\alias{qdODE_parallel}
\title{wrapper for qdODE_all in parallel version}
\usage{
qdODE_parallel(result, reduction = FALSE, thread = 2, maxit = 1000)
}
\arguments{
\item{result}{result from power_equation_fit}

\item{reduction}{use n/log(n) dimension reduction}

\item{thread}{scales for how many threads used}

\item{maxit}{scalar of Optim iteration setting}
}
\value{
list contain variable selection results and LOP parameters for every row
}
\description{
wrapper for qdODE_all in parallel version
}
