% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_plot_all}
\alias{qdODE_plot_all}
\title{plot all decompose plot}
\usage{
qdODE_plot_all(result)
}
\arguments{
\item{result}{list of qdODE parallel}
}
\value{
all effect curve decompose plot
}
\description{
plot all decompose plot
}
