% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_plot_base}
\alias{qdODE_plot_base}
\title{plot single decompose plot}
\usage{
qdODE_plot_base(result)
}
\arguments{
\item{result}{list of qdODE all}
}
\value{
effect curve decompose plot
}
\description{
plot single decompose plot
}
