% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODEmod}
\alias{qdODEmod}
\title{quasi-dynamic lotka volterra model}
\usage{
qdODEmod(Time, State, Pars, power_par)
}
\arguments{
\item{Time}{vector of time point}

\item{State}{vector of ODE initial state}

\item{Pars}{vector for unknown ODE parameters}

\item{power_par}{matrix of power equation parameters for dependent effect}
}
\value{
list used in ode function
}
\description{
quasi-dynamic lotka volterra model
}
