% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODEplot_convert}
\alias{qdODEplot_convert}
\title{convert qdODE results to plot data}
\usage{
qdODEplot_convert(result)
}
\arguments{
\item{result}{list of qdODE all}
}
\description{
convert qdODE results to plot data
}
