% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_functional_clustering.R
\name{bifun_clu_plot}
\alias{bifun_clu_plot}
\title{bifunctional clustering plot}
\usage{
bifun_clu_plot(
  result,
  best.k,
  label = 10,
  degree = 1/4,
  show.legend = FALSE,
  color1 = "#38E54D",
  color2 = "#FF8787"
)
}
\arguments{
\item{result}{list directly from bifun_clu_parallel function}

\item{best.k}{scale of BIC-determined cluster number}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{degree}{scalar control transparency degree}

\item{show.legend}{show legend or not}

\item{color1}{Hex Color Codes for first data}

\item{color2}{Hex Color Codes for second data}
}
\value{
functional clustering plot
}
\description{
bifunctional clustering plot
}
