% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R, R/bi_functional_clustering.R
\name{logsumexp}
\alias{logsumexp}
\title{calculate log-sum-exp values}
\usage{
logsumexp(x)

logsumexp(x)
}
\arguments{
\item{x}{}

\item{v}{numeric vector}
}
\value{
log-sum-exp values
}
\description{
calculate log-sum-exp values

Title
}
\examples{
logsumexp(c(100, 1000, 10000))
}
