% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mustard_microbe}
\alias{mustard_microbe}
\title{mustard microbe OTU data}
\format{
A data frame with 1557 rows and 176 column, contain first column as OTU ID:
}
\usage{
data(mustard_microbe)
}
\description{
Wagner, M. R. et al. Host genotype and age shape the leaf and root microbiomes of a wild perennial plant.
Nat. Commun. 7:12151 doi: 10.1038/ncomms12151 (2016)
This dataset is a subset of otuTable97, we select location = JAM, keep samples with both root
and leaf data, and then run data_cleaning first (set x = 50) to reduce size of this data.
Moreover, sample 8_1382 is removed for the outlier reason.
}
\keyword{datasets}
