% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reconstruction.R
\name{network_plot}
\alias{network_plot}
\title{generate network plot}
\usage{
network_plot(result, title = NULL, maxeffect = NULL, type = NULL)
}
\arguments{
\item{result}{list result from network_conversion}

\item{title}{text for plot title}

\item{maxeffect}{control edge size when compare networks}

\item{type}{select module effect or microbe effect}
}
\value{
network plot
}
\description{
generate network plot
}
