% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_plot_all}
\alias{qdODE_plot_all}
\title{plot all decompose plot}
\usage{
qdODE_plot_all(
  result,
  label = 10,
  show.legend = TRUE,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{result}{list of qdODE parallel}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{show.legend}{to show legend}

\item{nrow}{scalar for subplot row number}

\item{ncol}{scalar for subplot column number}
}
\value{
all effect curve decompose plot
}
\description{
plot all decompose plot
}
