% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solving.R
\name{qdODE_plot_base}
\alias{qdODE_plot_base}
\title{plot single decompose plot}
\usage{
qdODE_plot_base(result, label = 10, show.legend = TRUE)
}
\arguments{
\item{result}{list of qdODE all}

\item{label}{relabel x and y label due to log-transform, set 10 as default}

\item{show.legend}{to show legend}
}
\description{
plot single decompose plot
}
